Mise à jour gestion des ressources et gestion des informations : ajout du champ document, gestion des contacts

+++ app/Http/Controllers/ForumCommunicationController.php
@@ -49,9 +49,6 @@ class ForumCommunicationController extends Controller
             'image' => ['nullable', 'image', 'mimes:jpeg,png,jpg,gif'],
         ]);
 
-        // Récupérer l'utilisateur authentifié
-        //$user = auth()->user();
-
         // Créer une instance du modèle Forum_Communication avec les données validées
         $forum = new Forum_Communication([
             'titre' => $request->titre,
@@ -65,9 +62,6 @@ class ForumCommunicationController extends Controller
             $imageFile->move(public_path('images'), $filename);
             $forum->image = $filename;
         }
-
-        // Assigner l'utilisateur authentifié comme créateur du forum
-       // $forum->user_id = $user->id;
 
         // Sauvegarder le forum
         $forum->save();
@@ -130,21 +124,21 @@ class ForumCommunicationController extends Controller
     public function update(Request $request, $id)
     {
         try {
-            // $request->validate([
-            //     'titre' => ['required', 'string'],
-            //     'texte' => ['required', 'string'],
-            //     'image' => ['nullable', 'image', 'mimes:jpeg,png,jpg,gif'],
-            // ]);
+            $request->validate([
+                'titre' => ['required', 'string'],
+                'texte' => ['required', 'string'],
+                'image' => ['nullable', 'image', 'mimes:jpeg,png,jpg,gif'],
+            ]);
     
             $forum = Forum_Communication::findOrFail($id);
     
-            // // Vérifier si l'utilisateur actuel est l'auteur du forum
-            // if ($forum->user_id !== auth()->id()) {
-            //     return response()->json([
-            //         'code_valide' => 403,
-            //         'message' => 'Vous n\'avez pas la permission de modifier ce forum.',
-            //     ], 403);
-            // }
+            // Vérifier si l'utilisateur actuel est l'auteur du forum
+            if ($forum->user_id !== auth()->id()) {
+                return response()->json([
+                    'code_valide' => 403,
+                    'message' => 'Vous n\'avez pas la permission de modifier ce forum.',
+                ], 403);
+            }
     
             // Mettre à jour les attributs du forum
             $forum->titre = $request->titre;
@@ -180,41 +174,22 @@ class ForumCommunicationController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Forum_Communication $id)
+    public function destroy($id)
     {
-    //     try {
-    //         // Vérifier si l'utilisateur est un administrateur
-    //         if (Auth::user()->role !== 'admin') {
-    //             return response()->json([
-    //                 'message' => 'Vous n\'avez pas les droits pour supprimer un forum.',
-    //             ], 403); // Statut HTTP 403: Accès interdit
-    //         }
 
-    //         $forum = Forum_Communication::findOrFail($id);
-    //         $forum->delete();
-
-    //         return response()->json([
-    //             'message' => 'Forum supprimé avec succès.',
-    //         ], 200);
-    //     } catch (\Exception $e) {
-    //         return response()->json([
-    //             'message' => 'Une erreur s\'est produite lors de la suppression du forum.',
-    //             'error' => $e->getMessage(),
-    //         ], 500);
-    //     }
     try {
         // Récupérer le forum
         $forum = Forum_Communication::findOrFail($id);
 
-        // Vérifier si l'utilisateur actuel est l'auteur du forum ou s'il a le rôle d'administrateur
-        // if ($forum->user_id !== auth()->id() && auth()->user()->role !== 'admin') {
-        //     return response()->json([
-        //         'code_valide' => 403,
-        //         'message' => 'Vous n\'avez pas la permission de supprimer ce forum.',
-        //     ], 403);
-        // }
+        //Vérifier si l'utilisateur actuel est l'auteur du forum ou s'il a le rôle d'administrateur
+        if ($forum->user_id !== auth()->id() && auth()->user()->role !== 'admin') {
+            return response()->json([
+                'code_valide' => 403, // Statut HTTP 403: Accès interdit
+                'message' => 'Vous n\'avez pas la permission de supprimer ce forum.',
+            ], 403);
+        }
 
-        // Supprimer le forum
+        //Supprimer le forum
         $forum->delete();
 
         return response()->json([
